If cursor position same as last point move cursor (instead of failing)
Failing was happening to me often in testing, so I expect it to happen to more to regular users. Also reported message upon failing was incorrect.

There seems to be a bug with `context.window.cursor_wrap` where it jumps cursor on y axis every time, and should be reportable to Blender.