Preference for choosing order when applying modifiers
In previous versions of Bool Tool object was converted to mesh when using auto-boolean, or using "Apply All" operator. This feature was requested back.

This commit introduces three ways of choosing how to apply modifiers and choice is in preferences:
1. Apply all modifiers (convert to mesh, like before)
2. Only apply boolean modifiers (like current version)
3. Apply everything before last boolean modifier.

For third option there is new function for getting all modifiers before last boolean. This preference applies to auto-boolean and "Apply All Booleans" operator. It doesn't apply to applying individual cutters, or Carver tool.