Handle objects with shape keys
Objects with shape keys were never accounted for before, and resulted in various long python errors. Shape keys are only problem when boolean modifier has to be applied, so this commit adds polls for each of those instances to either cancel the operation alltogether if object has shape keys, or remove it from list.

This is very context dependent and various things can dictate what happens. For example, if applying all modifiers by converting it to mesh, shape keys are applied too.