Pin boolean modifiers during brush and auto booleans
Pinning boolean modifiers was added when porting Carver tool from "Carver" add-on and integrated into `add_boolean_modifier` function. It is now easy to integrate the option for brush and auto booleans as well. Added "Pin Boolean Modifiers" in preferences for it (it doesn't affect Carver tool, since it has it's own property for it)

This, alongside with `apply_order` preference introduced earlier create new workflow: When pinning is enabled and apply order is "Apply Booleans & Everything Before" performing auto-boolean will only apply modifiers that are before final boolean (if there is other boolean). This was not possible before and choice was between applying everything or only boolean.