Fix (unreported): Cutter jumps in viewport when removing/changing parent
When cutter parent was removed during "Remove Cutter" or when parent was changed to something else automatically it would jump to original position, because it was parented with inverse matrix.

This could be avoided entirely if I parent without inverse matrix (but still keep transformations), but anyway good to have: extracted function for changing parent (which can be removed by passing None as parent arg) that copies matrix before and applies it to cutter after changing parent.