Fix: Carver: Object origin not centered on cutters
Because I removed `bpy.ops` method of setting object origin it was now always staying at world center. This commit adds a function for setting object origin at given position by shifting vertices in opposite, and for position it uses center of the bounding box, which should be enough for cutters.

Not sure about the performance of this vs `bpy.ops` method, but will likely have no impact. If that turns out false I'll revisit this.