Carver: Purge cutters when using "Remove All Cutters" operator
While the behavior is different from using brush booleans, I think this is justified, because:
1. Cutters are created by the operator, now by the user, so they have no meaning in the scene if they're not cutting anything.
2. If "Hide Cutter" is enabled and user doesn't see cutters removing all of them will simply remove modifiers, but keep the objects hidden. User may never realize they're still there, and deleting entire "boolean_cutters" collection isn't an option if you have active cutters.

This adds new `booleans.carver` property which is set for cutters that are carving the object. Using name here would be too fickle since "cutter" is too common a name and user might set that name to custom objects themselves, and besides this property might come in handy in future.