Fix: Carver: Error when initializing operator with center origin
If operator is called with origin to center (with Alt) and `PRESS` by pressing, it returns error because first point isn't initiated before cursor is moved. This commit disallows it and instead changes shortcut to `CLICK_DRAG` to that only dragging can start shape creation.

This also frees up alt click to select objects.