Fix #26: Python error in `modifier_poll` without active object
After recent changes to both this poll and dbl-click select feature in general this will not appear anymore, but anyway changed a bit how the poll worked and turned it into `active_modifier_poll`. I was thinking about using this outside this feature, but so far have no uses for it, so might as well make it even safer.