Carver: Write modifier keys in status bar instead of toolbar
Text is the same, still can't fully do icons because operator `self` isn't accessible in function that sets status bar `draw`, but at least it's in the right place now.

In future with very intense hacks this can be made to look even better, but splitting operator is necessary first.