Fix (unreported): Carver: Python error when canvas has no geometry
If canvas had no geometry for some reason bmesh still counted it as having one edge, tried working on it, and failed, which threw python errors. And when it didn't, it added modifiers on empty object, which is unnecessary.

Added check for object dimensions in selection fallback. If it's 0, 0, 0, which most likely means that object has no geometry, it will filter out the object.