Allow destructive operators on objects with shape keys
New, main method for applying modifiers also takes shape keys into account, because it uses an evaluated object. Which means after new mesh is created it is identical to the shape key result, so none of the shape keys do anything anymore.

Only thing we can do here is to simply remove shape keys and consider them applied as well.

Previous to this none of the destructive operators allowed working on objects with shape keys (except edge cases). That in itself was not ideal, so current method of applying them is preferred.

This commit allows executing destructive operators on objects with shape keys & applies them in fallback mode of the `apply_modifiers` function as well, so that both implementations match.