Simplify & optimize applying modifiers and apply order
This part was so complicated and messy (mostly because of its combinations with pin preference) that it's difficult to describe what exactly was wrong. Some of the options were just not working. Function for applying modifiers was also called for each cutter, which added unnecessary executions with new method. All of this is changed and simplified now.

Other changes:
- Refactored `add_boolean_modifier` function. It now returns modifier, instead of None. Apply argument was also removed, as it's not used, applying is a separate function.
- Preferences "Apply Order" and "Pin Boolean Modifiers" now applies to slice objects as well.