Confirmation pop-up for objects with shape keys
In one of the previous commits destructive operators were allowed on objects that have shape keys. This commit adds confirmation window for them, similar to one for objects with instanced data (in fact that one was generalized in a function that is reused here).

Basically, if canvas object has shape keys, operators will ask user if they're sure they want to apply modifiers, instead of failing.